Comments
Almost everything is commented out, but the loops were necessary to find the keys we needed.  The code on line 8 is where I used those keys to find the correct data.  The error reporting should be turned on at the top, and the url will need to be uncommented when this is run locally.  This was one of my first php files, so I've left some code bits in here for later reference.